chore: Add ruff, mypy, coverage, dependabot and CI jobs
Sets up the tooling the repository was missing: ruff (lint + format) over
everything, mypy in strict mode over the Home-Assistant-free modules only —
those are already fully annotated, and the rest would need the whole framework
in the type check for very little gain.

The existing code is reformatted by ruff format in the same commit; the only
non-mechanical changes are two nested ifs merged and a redundant int(round()).

CI gains a lint job, a card job (vitest and a build smoke test, which so far
only ran during a release) and a placeholder job for the Home-Assistant tests
that follow. The release workflow now checks card/package.json against the tag
as well — the card carries its own version and is bumped in the same step.

The coverage floor is 40 % for now, deliberately below the current 44 %: it is
there to catch a test file falling out of collection, not to be a target. It
goes up once the Home-Assistant tests land.

+++ custom_components/crowdsec/timewindow.py
@@ -46,7 +46,7 @@ def parse_duration(text: str) -> int | None:
         return None
     if amount <= 0:
         return None
-    return max(1, int(round(amount * factor)))
+    return max(1, round(amount * factor))
 
 
 def window_params(window: Window) -> dict[str, str]:
